Added template parameters to almost all things in the analysis-synthesis pipeline; added Residual class as a vector of momentary spectra; exact residual synthesis done

+++ plugins/Diginstrument/common/Qt/DimensionField.hpp
@@ -55,7 +55,7 @@ class DimensionField : public QWidget
       layout->setSpacing(0);
     }
 
-    std::pair<std::string, double> getCoordinate() const
+    std::pair<std::string, float> getCoordinate() const
     {
       return std::make_pair(label->text().toStdString(), value->text().toDouble());
     }
