+++ plugins/Diginstrument/player/DiginstrumentView.cpp
@@ -77,7 +77,7 @@ void DiginstrumentView::showInstumentVisualization()
   //visualization->adjustSize();
 }
 
-void DiginstrumentView::updateVisualizationData(float minTime, float maxTime, float minFreq, float maxFreq, int timeSamples, int freqSamples, std::vector<double> coordinates)
+void DiginstrumentView::updateVisualizationData(float minTime, float maxTime, float minFreq, float maxFreq, int timeSamples, int freqSamples, std::vector<float> coordinates)
 {
   visualization->setSurfaceData(castModel<DiginstrumentPlugin>()->getInstrumentSurfaceData(minTime/1000.0f,maxTime/1000.0f,minFreq,maxFreq,timeSamples,freqSamples, coordinates));
   visualization->removeCustomItems();
